Move ActionMenu.vue to utils
